add menu button to mobile
