+++ packages/frontend/components/ui/nav-icon-button.js
@@ -1,22 +1,24 @@
 import React from 'react';
-import styled from '@emotion/styled';
 import PropTypes from 'prop-types';
-import { IconButton, Badge } from '@chakra-ui/core';
+import { IconButton, Badge, Box } from '@chakra-ui/core';
 
-const Wrapper = styled.div`
-  position: relative;
-  margin-right: 2rem;
-`;
-
-function NavButton({ icon, label, badge, children }) {
+function NavButton({
+  icon,
+  label,
+  badge,
+  children,
+  wrapperProps,
+  ...otherProps
+}) {
   return (
-    <Wrapper>
+    <Box position="relative" ml={4} {...wrapperProps}>
       <IconButton
         aria-label={label}
         icon={icon}
         fontSize="22px"
         variant="ghost"
         color="paragraph"
+        {...otherProps}
       />
       {badge && (
         <Badge
@@ -39,7 +41,7 @@ function NavButton({ icon, label, badge, children }) {
         </Badge>
       )}
       {children}
-    </Wrapper>
+    </Box>
   );
 }
 
@@ -48,6 +50,8 @@ NavButton.propTypes = {
     .isRequired,
   label: PropTypes.string.isRequired,
   badge: PropTypes.number,
+  // eslint-disable-next-line react/forbid-prop-types
+  wrapperProps: PropTypes.object,
   children: PropTypes.node,
 };
