+++ packages/frontend/pages/account/signup.js
@@ -3,7 +3,7 @@ import { Link } from '@chakra-ui/core';
 import NextLink from 'next/link';
 import SignupForm from '../../components/forms/signup-form';
 import WaitingForYou from '../../components/illustrations/waiting-for-you';
-import Header from '../../components/header';
+import Header from '../../components/ui/header';
 import Container from '../../components/ui/container';
 import AccountFormTemplate from '../../components/ui/templates/account-form-template';
