+++ packages/frontend/pages/index.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Header from '../components/header';
+import Header from '../components/ui/header';
 
 function Home() {
   return (
